Pause animation when it ends.
The slider menu animation is intended to be controlled with touch
events. When the animation is being controlled that way, it must be
paused, otherwise it will be trying to complete at the same time that
the user sets its current time, leading to visual glitches.

I was ensuring that the animation was paused in each touchmove event,
but now the animation is always paused on its `onfinish` callback, so
I can avoid doing that work over and over in each touchmove event.